+++ src/main/java/io/github/astrapi69/entity/id/generator/BigIntegerIdGenerator.java
@@ -24,40 +24,57 @@
  */
 package io.github.astrapi69.entity.id.generator;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.math.BigInteger;
 
+import lombok.NonNull;
 import io.github.astrapi69.data.identifiable.IdGenerator;
 
 /**
- * The class {@link SystemTimeIdGenerator} is an base implementation of {@link IdGenerator}
- * interface
+ * The class {@link BigIntegerIdGenerator} is an implementation of {@link IdGenerator} interface
+ * with id type of {@link BigInteger} object
  */
-public class SystemTimeIdGenerator implements IdGenerator<Integer>
+public class BigIntegerIdGenerator implements IdGenerator<BigInteger>
 {
 
 	/**
-	 * The instance.
+	 * The default instance with an initial value of 0
 	 */
-	private static final SystemTimeIdGenerator instance = new SystemTimeIdGenerator();
+	private static final BigIntegerIdGenerator instance = new BigIntegerIdGenerator(
+		BigInteger.ZERO);
+
 	/**
 	 * The atomic id.
 	 */
-	private final AtomicInteger atomicId;
+	private final AtomicBigInteger atomicId;
 
 	/**
-	 * Instantiates a new system time id generator.
+	 * Instantiates a new {@link BigIntegerIdGenerator}
+	 *
+	 * @param initialValue
+	 *            the initial value for the generator
 	 */
-	private SystemTimeIdGenerator()
+	private BigIntegerIdGenerator(final @NonNull BigInteger initialValue)
 	{
-		atomicId = new AtomicInteger((int)System.currentTimeMillis());
+		atomicId = new AtomicBigInteger(initialValue);
 	}
 
 	/**
-	 * Gets the single instance of SystemTimeIdGenerator.
+	 * Factory method for create a new custom {@link BigIntegerIdGenerator} with an initial value
 	 *
-	 * @return single instance of SystemTimeIdGenerator
+	 * @param initialValue
+	 *            the initial value for the generator
 	 */
-	public static SystemTimeIdGenerator getInstance()
+	public static BigIntegerIdGenerator of(final @NonNull BigInteger initialValue)
+	{
+		return new BigIntegerIdGenerator(initialValue);
+	}
+
+	/**
+	 * Gets the single instance of {@link BigIntegerIdGenerator}.
+	 *
+	 * @return single instance of {@link BigIntegerIdGenerator}
+	 */
+	public static BigIntegerIdGenerator getInstance()
 	{
 		return instance;
 	}
@@ -66,13 +83,8 @@ public class SystemTimeIdGenerator implements IdGenerator<Integer>
 	 * {@inheritDoc}
 	 */
 	@Override
-	public Integer getNextId()
+	public BigInteger getNextId()
 	{
-		int nextId = atomicId.getAndIncrement();
-		if (nextId < 0)
-		{
-			nextId *= -1;
-		}
-		return nextId;
+		return atomicId.getAndIncrement();
 	}
 }