+++ src/main/java/io/github/astrapi69/entity/id/generator/IdGenerator.java
@@ -25,14 +25,17 @@
 package io.github.astrapi69.entity.id.generator;
 
 /**
- * The Interface IdGenerator.
+ * The interface {@link IdGenerator} provides a method for get the next id.
+ *
+ * @deprecated use instead the same name interface from module data-api
  */
+@Deprecated
 public interface IdGenerator
 {
 
 	/**
 	 * Gets the next id.
-	 * 
+	 *
 	 * @return the next id
 	 */
 	int getNextId();