+++ src/test/java/io/github/astrapi69/entity/id/generator/LongIdGeneratorTest.java
@@ -24,44 +24,31 @@
  */
 package io.github.astrapi69.entity.id.generator;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
-import org.meanbean.test.BeanTester;
 
-/**
- * The unit test class for the class {@link SystemTimeIdGenerator}.
- */
-public class SystemTimeIdGeneratorTest
+public class LongIdGeneratorTest
 {
 
 	/**
-	 * Test method for {@link SystemTimeIdGenerator#getNextId()}
+	 * Test method for {@link LongIdGenerator#getNextId()}
 	 */
 	@Test
 	void getNextId()
 	{
-		final List<Integer> nextIds = new ArrayList<>();
+		final List<Long> nextIds = new ArrayList<>();
 		for (int i = 0; i < 1000; i++)
 		{
-			final int nextId = SystemTimeIdGenerator.getInstance().getNextId();
+			final Long nextId = LongIdGenerator.getInstance().getNextId();
 			assertFalse(nextIds.contains(nextId),
 				"Next id" + nextId + " should not generated twice");
 			nextIds.add(nextId);
 		}
-		System.out.println(nextIds);
-	}
-
-	/**
-	 * Test method for {@link SystemTimeIdGenerator}
-	 */
-	@Test
-	public void testWithBeanTester()
-	{
-		final BeanTester beanTester = new BeanTester();
-		beanTester.testBean(SystemTimeIdGenerator.class);
+		assertEquals(nextIds.size(), 1000);
 	}
 }